Influence conflict modification with string flags
To make mappings more meaningful it is now possible to request 'append'
behavior with a string flag of the same name.

It is also possible to modify a conflict from any location in the right
file, by specifying the 'jumpto' flag.  By default it is still an error
if not within the conflict range to give the user more certainty which
conflict is being edited.