Provide default mappings in locked mode
Because the buffer is unmodifiable when a conflict is locked, it is
possible to reuse some major normal mode commands to skip through
slides.  However, there is no actual backup and restore mechanism in
case there are buffer-local mappings on the keys.  The intention is to
get the user going quickly so that the usefulness of this plugin can be
evaluated.

It can be turned off by setting the variable
g:conflictslides_handle_locked_mappings to zero.