Release the lock in the owning buffer
To release a lock in a buffer that is not the current one, switch to it
and execute the whole operation in the buffer with the lock.  To not
disturb any diff settings, operate in another tab and close it when
finished.

This fixes the unmapping errors when trying to release a lock from
another buffer.