Add display helper functions
Conflict slides become really useful when there are mappings that (in
addition) turn on/off the diff in selected windows.  For example, you
might want to insert the 'base' slide and turn off &diff in ours or
theirs to see a 2-way diff for the corresponding other one.  Or when you
insert 'our' slide, it's probably enough to see the diff between dest
and theirs.  To accomplish that there are a few helper functions that
you can incorporate into your mappings.  Hint: the 1,2,3 keys with some
prefix are ideal for toggling in my opinion.

Note: You can maneuver yourself into a situation where there is no
difference in two active diff-windows and so the whole conflict range is
folded away.  In that case you can no longer change the conflict without
the 'jumpto' flag.  Just turn off folding (eg. with 'zn') or have some
mapping handy that turns on diff in all windows.

You'll soon notice terrible jolting when switching &diff while the
folding is on.  I think, it makes sense to turn the folding off in the
post-lock callback and turn it back on in selected mappings (for example
plain unlocking).  This commit adds a function that you can use to turn
the folding on/off or toggle it.