Search conflict marker only at start of line when moving cursor
This fixes some cases where you try to lock the next conflict and end up
with a "Not within conflict range" error because a conflict marker
('<<<<<<<') appears somewhere in the source code.  It will still happen
if such a marker appears at the start of a line for some reason.